refactor(blog): extract OwnerActions, scope sticky to md+, add tests
- Move OwnerActions (and the OwnerIcon helper) out of PostDetail.jsx
  into its own component, so the visibility check can be tested in
  isolation.
- Replace the magic numbers (top: 68, zIndex: 9) with two CSS custom
  properties — --header-height and --z-owner-actions — declared in
  styles/index.css. The TOC sticky offset on lg+ now reads from the
  same variable so a future header-height change only needs one edit.
- Make the bar sticky only from the md breakpoint (>= 768px). On
  mobile the toolbar is now position: static so it no longer eats
  vertical space on small viewports.
- Add Vitest + Testing Library coverage for visibility (anonymous,
  non-superuser owner, superuser non-owner, superuser owner) and for
  the responsive sticky utility classes.

Refs #258
